+++ src/components/messegeList/messageList.js
@@ -1,0 +1,7 @@
+import Message from './../message/message';
+
+const MessageList = ({ messageList }) => {
+    return messageList.map(el => <Message key={el.author + Math.random()} messageText={el.text} author={el.author} />)
+}
+
+export default MessageList;
