+++ src/utils/constants.js
@@ -1,0 +1,4 @@
+export const AUTHORS = {
+    AUTHOR_BOT: "Bot",
+    AUTHOR_ME: "Me", 
+}
